Completely rework token validation
We've now switched to always using JWKS for token validation, as that is
standardized. Vault OIDC for example doesn't have a 'public key' that
can be used for validation and only supports JWKS.

The JWKS of the token issuer will be grabbed and cached. The system will
detect when a JWKS is no longer valid and invalidate + refresh the
cache.